refactor(config): fold ClassMethods mixin into Config class
